Fixed bug in sprite code that caused engine crashes, introduced in 2851.
Also, the fillcolor parameter of the drawmethod command can now be in the format "R_G_B".  For example, a yellow fillcolor (red 255, green 255, blue 0) can be specified as 255_255_0.  Unlike the old format for fillcolor (which still works for backwards compatibility), the new format isn't dependent on color depth and doesn't require a complicated formula to calculate.